safari-534.53-branch/LayoutTests/fast/preloader/input.html

<body>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpResourceResponseMIMETypes();
}
</script>
This test requires DumpRenderTree to see the log of what resources are loaded.
<script src=resources/non-existant.js></script>
<script>document.write("<plaintext>");</script>
<input src=resources/fail.png type=BUTTON>
<input src=resources/image1.png type=IMAGE>
