LayoutTests/fast/preloader/input-expected.txt

This test requires DumpRenderTree to see the log of what resources are loaded.

<input src=resources/fail.png type=BUTTON>
<input src=resources/image1.png type=IMAGE>

input.html has MIME type text/html
image1.png has MIME type image/png
